pyproject.toml
[project]
name = "skating-biomechanics-ml"
version = "0.1.0"
description = "ML-based figure skating biomechanics analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30",
    "bcrypt>=4.0,<4.1",
    "email-validator>=2.3.0",
    "rtmlib>=0.0.7",            # RTMPose pose estimation (HALPE26, 26kp with feet)
    "onnxruntime-gpu>=1.24.4",      # ONNX Runtime with CUDA backend for rtmlib
    "opencv-python>=4.10.0",
    "numpy>=2.0.0",
    # torch is optional — install with: uv add torch --extra torch-models
    "dtw-python>=1.5.0",
    "scipy>=1.14.0",
    "pillow>=12.1.1",
    "deep-sort-realtime>=1.3.2",
    "av>=17.0.0",
    "trimesh>=4.11.5",
    "pygltflib>=1.16.5",
    "fastapi>=0.135.2",
    "passlib[bcrypt]>=1.7",
    "PyJWT>=2.9",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.22",
    "redis>=5.0.0",
    "sse-starlette>=3.3.4",
    "sqlalchemy[asyncio]>=2.0",
    "structlog>=25.5.0",
    "ultralytics>=8.0.0",
    "uvicorn[standard]>=0.42.0",
    "alembic>=1.18.4",
    "arq>=0.27.0",
    "boto3>=1.42.83",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app", "ml/skating_ml"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "basedpyright>=1.37.0",
    "vulture>=2.0",
    "jupyter>=1.0.0",
    "onnxscript>=0.6.2",
    "pytest-timeout>=2.4.0",
    "structlog>=25.5.0",
    "bandit>=1.9.4",
    "pytest-asyncio>=1.3.0",
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
deepsort = ["deep-sort-realtime>=1.3.2"]

[project.scripts]
analyze = "cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["Sports2D/", "TCPFormer/"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented out code)
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "PLR0912", # too many branches (CLI scripts with many flags)
    "PLR0915", # too many statements (long pipeline functions)
    "PLR2004", # magic value comparison
    "PLC0415", # import not at top-level (lazy imports for heavy deps)
    "SIM108", # ternary instead of if-else
    "ARG002", # unused method argument (common in callback interfaces)
    "RUF001", # ambiguous Cyrillic chars in strings (Russian UI text)
    "RUF002", # ambiguous Cyrillic chars in docstrings (Russian docs)
    "RUF003", # ambiguous Cyrillic chars in comments (Russian comments)
]
unfixable = [
    "ERA",    # don't auto-remove commented out code
]

[tool.ruff.lint.pylint]
max-args = 12
max-branches = 35
max-returns = 8
max-statements = 80

[tool.ruff.lint.isort]
known-first-party = ["backend", "skating_ml"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG", "PLR2004", "ERA", "SIM117"]
"__init__.py" = ["F401"]  # unused imports
"scripts/visualize_with_skeleton.py" = ["ARG001"]
"ml/skating_ml/analysis/rules/*.py" = []
"ml/skating_ml/visualization/**/*.py" = []
"ml/skating_ml/cli.py" = []
"ml/**/*.py" = ["ARG001", "F841"]  # unused args in callback-heavy code
"backend/**/*.py" = ["ARG001", "F841"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.basedpyright]
typeCheckingMode = "standard"
reportMissingTypeStubs = false
reportUnknownMemberType = "warning"
reportUnknownVariableType = "warning"
reportUnknownParameterType = "warning"
reportUnknownArgumentType = "warning"
reportGeneralTypeIssues = "warning"
reportExplicitAny = "warning"
reportAny = "warning"
reportUnusedCallResult = "warning"
reportUnusedVariable = "warning"
pythonVersion = "3.11"
exclude = [
    "Sports2D/",
    "TCPFormer/",
]

[tool.vulture]
min_confidence = 80
paths = ["backend/app", "ml/skating_ml", "backend/tests", "ml/tests"]
exclude = ["__init__.py"]
make_whitelist = false

[tool.bandit]
skips = ["B101", "B608"]
exclude_dirs = ["Sports2D/", "TCPFormer/", "tests/"]

[tool.pytest.ini_options]
testpaths = ["backend/tests", "ml/tests"]
pythonpath = ["backend", "ml"]
addopts = [
    "--cov=backend/app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
    "--timeout=60",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["backend/app", "ml/skating_ml"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
